+++ DesignToolbox/DesignToolbox/Pages/Components/Indicators/Badge/BadgeCount/BadgeCountElement.swift
@@ -1,0 +1,31 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import OUDSSwiftUI
+import SwiftUI
+
+struct BadgeCountElement: DesignToolboxElement {
+    let name: String
+    let illustration: AnyView
+    let pageDescription: AnyView
+
+    init() {
+        name = "app_components_badge_count_tech".localized()
+        illustration = AnyView(BadgeIllustration())
+        pageDescription = AnyView(DesignToolboxElementPage(
+            name: name,
+            description: "app_components_badge_count_description_text",
+            version: OUDSVersions.componentBadgeCountVersion,
+            demoScreen: AnyView(BadgeCountPage())))
+    }
+}