+++ DesignToolbox/DesignToolbox/Pages/Components/Indicators/Badge/BadgeElements.swift
@@ -1,0 +1,44 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import OUDSSwiftUI
+import SwiftUI
+
+struct BadgeElements: DesignToolboxElement {
+    let name: String
+    let illustration: AnyView
+    let pageDescription: AnyView
+
+    init() {
+        let variants: [DesignToolboxElement] = [
+            BadgeStandardElement(),
+            BadgeCountElement(),
+            BadgeIconElement(),
+        ]
+
+        name = "app_components_badge_tech".localized()
+        illustration = AnyView(BadgeIllustration())
+        pageDescription = AnyView(DesignToolboxElementPage(
+            name: name,
+            illustration: illustration,
+            description: "app_components_badge_description_text",
+            demoScreen: AnyView(DesignToolboxVariantElement(elements: variants))))
+    }
+}
+
+struct BadgeIllustration: View {
+
+    var body: some View {
+        OUDSBadgeCount(1, accessibilityLabel: "", status: .negative, size: .medium)
+    }
+}